Add ESLint extension
Really loving your extension pack.

I just noticed that the ESLint extension is not included in this pack
and I use it all the time, so I thought this could be a good idea
to include it in this pack.

Feel free to improve the description as it's quite rudimentary for now.